Fixed MSVC warning.

+++ src/opal/endpoint.cxx
@@ -209,7 +209,7 @@ PStringArray OpalEndPoint::GetDefaultListeners() const
     WORD port = defaultSignalPort;
     PINDEX pos = t.Find(':');
     if (pos != P_MAX_INDEX) {
-      port = t.Mid(pos+1).AsUnsigned();
+      port = (WORD)t.Mid(pos+1).AsUnsigned();
       t = t.Left(pos);
     }
     PString listenerAddress = t + '*';
